+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Temporary persistent volumes generation and removal for pipelines
+
 ## [0.0.4] - 2021-03-10
 
 ## [0.0.3] - 2021-03-09