+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Creates mlflow experiment on pipeline start if it does not exist
 -   Temporary persistent volumes generation and removal for pipelines
+-   `upload-pipeline` command that generates DAG to Airflow directory 
 
 ## [0.0.4] - 2021-03-10